change configuration format and 0.2.0 release
Chanes the configuration format to be more consistent and easier to
generate a parser for. This version still uses a hand-written parser
but this change is in preperation for a proc-macro generated
configuration parser. That should hopefully not affect the format anymore.

The change is rather trivial to fix but will break every configuration
file containing at least one task. In 0.1.x tasks were defined using
table values in the global table with otherwise-unused name:

```
[task_name]
cmd = "..."
```

From this version onwards tasks are defined using an array of tables.
The example above would be written as:

```
[[task]]
name = "task_name"
cmd = "..."
```